backend/requirements.txt
django
djangorestframework
textblob
